First commit
Basic modular functionality is there. Config is read from a json file and pages with modules are created from there.
Modules have defined imports and exports and can reactively import variables from modules that export them.
Using R6 object-oriented programming, everything is a lot cleaner and easier to manage. A generic Module class contains the necessary boilerplate code and subclasses contain only the code that is specific to that module.
A destroy method is also available in Module that tries to remove and destroy as much as possible.